+++ src/main/java/com/researchspace/model/preference/Preference.java
@@ -116,7 +116,10 @@ public enum Preference {
 	CURRENT_WORKSPACE_VIEW_MODE("LIST_VIEW", SettingsType.STRING, PreferenceCategory.UI,
 			"Whether Workspace page opens up as list view or tree view initially"),
 	DELETED_RECORDS_RESULTS_PER_PAGE("10", SettingsType.NUMBER, PreferenceCategory.UI,
-			"Number of deleted items per page");
+			"Number of deleted items per page"),
+
+	/** Similar to UI_CLIENT_SETTINGS, but for storing arbitrary data in json key-value format  */
+	UI_JSON_SETTINGS("", SettingsType.TEXT, PreferenceCategory.UI, "UI client-side settings, in json format");
 
 	private String defaultValue;
 	private SettingsType prefType;